+++ widgets/src/grid_view.rs
@@ -2,7 +2,7 @@ use ribir_core::prelude::*;
 
 use crate::layout::Direction;
 
-#[derive(MultiChild, Query)]
+#[derive(MultiChild)]
 pub struct GridView {
   axis_dir: Direction,
   cross_axis_cnt: u32,